// src/shared/api/@generated/models/follower-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface FollowerResponse {
    'id'?: number;
    'nickname': string;
    'profileImgNumber'?: number;
    'handle': string;
}